29092022/1211 - mutex destruction not working yet. everything elese seems to be ok

+++ ft_atoi.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	unsigned int	i;
 	long long		res;
